requirements.txt

Flask==2.0.3
openai==0.22.0
python-dotenv